backend/requirements.txt
fastapi
uvicorn
PyMuPDF==1.20.2
google-cloud-documentai==2.24.1
google-cloud-storage==2.15.0
google-cloud-documentai-toolbox==0.13.3a0
paddlepaddle==2.6.1
paddleocr==2.7.0.3
transformers==4.38.2
langchain==0.1.12
langchain-text-splitters
langchain-google-vertexai==0.1.1
torch==2.2.1
torchvision==0.17.1
google-generativeai==0.4.1
langchain-google-genai
qdrant-client
ragas